backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
pgvector>=0.3.6
alembic>=1.14.0
google-adk>=1.5.0
google-genai>=1.14.0
trafilatura>=2.0.0
readability-lxml>=0.8.1
pydantic-settings>=2.7.0
httpx>=0.28.1
python-multipart>=0.0.20
sse-starlette>=2.2.1
